refactor(deferred-task): move date validation utilities to date-util module and optimize task deferral check
