+++ lib/utils.ts
@@ -31,34 +31,3 @@ export function hexToRgba(hex: string, alpha: number) {
   const b = num & 255
   return `rgba(${r},${g},${b},${alpha})`
 }
-
-export const isDateValidForDefer = (date: Date, dueDate?: string): boolean => {
-  const today = new Date()
-  today.setHours(0, 0, 0, 0)
-
-  if (date <= today) return false
-
-  if (dueDate) {
-    const taskDueDate = new Date(dueDate)
-    taskDueDate.setHours(0, 0, 0, 0)
-    if (date >= taskDueDate) return false
-  }
-
-  return true
-}
-
-export const hasValidDeferDates = (dueDate?: string): boolean => {
-  const today = new Date()
-  today.setHours(0, 0, 0, 0)
-
-  if (!dueDate) {
-    return true
-  }
-
-  const taskDueDate = new Date(dueDate)
-  taskDueDate.setHours(0, 0, 0, 0)
-
-  const tomorrow = new Date(today)
-  tomorrow.setDate(tomorrow.getDate() + 1)
-  return tomorrow < taskDueDate
-}
